+++ src/main/commands/pdfGenerator/pdfGenerator.js
@@ -13,55 +13,47 @@ const getCVLatex = require('./templates/template2');
  * @returns {Promise}
  */
 const pdfGenerator = async (latexString, type, id) => {
-
   let latexPdf;
-  try {
-    latexPdf = await latex(latexString);
-  } catch (error) {
-    throw error;
+
+  latexPdf = await latex(latexString);
+
+  // Check for existing files and delete existing file
+  const pdfFiles = await fsPromises.readdir(
+    path.join(app.getPath('userData'), `output_files/${type}_pdf`)
+  );
+  const existingPdfFile = pdfFiles.find((file) => {
+    const pattern = new RegExp(type + '_' + id);
+    return pattern.test(file);
+  });
+  if (existingPdfFile !== undefined) {
+    await fsPromises.unlink(
+      path.join(
+        app.getPath('userData'),
+        `output_files/${type}_pdf`,
+        existingPdfFile
+      )
+    );
   }
 
-  try {
-    // Check for existing files and delete existing file
-    const pdfFiles = await fsPromises.readdir(
-      path.join(app.getPath('userData'), `output_files/${type}_pdf`)
+  const texFiles = await fsPromises.readdir(
+    path.join(app.getPath('userData'), `output_files/${type}_tex`)
+  );
+  const existingTexFile = texFiles.find((file) => {
+    const pattern = new RegExp(type + '_' + id);
+    return pattern.test(file);
+  });
+  if (existingTexFile !== undefined) {
+    await fsPromises.unlink(
+      path.join(
+        app.getPath('userData'),
+        `output_files/${type}_tex`,
+        existingTexFile
+      )
     );
-    const existingPdfFile = pdfFiles.find((file) => {
-      const pattern = new RegExp(type + '_' + id);
-      return pattern.test(file);
-    });
-    if (existingPdfFile !== undefined) {
-      await fsPromises.unlink(
-        path.join(
-          app.getPath('userData'),
-          `output_files/${type}_pdf`,
-          existingPdfFile
-        )
-      );
-    }
-
-    const texFiles = await fsPromises.readdir(
-      path.join(app.getPath('userData'), `output_files/${type}_tex`)
-    );
-    const existingTexFile = texFiles.find((file) => {
-      const pattern = new RegExp(type + '_' + id);
-      return pattern.test(file);
-    });
-    if (existingTexFile !== undefined) {
-      await fsPromises.unlink(
-        path.join(
-          app.getPath('userData'),
-          `output_files/${type}_tex`,
-          existingTexFile
-        )
-      );
-    }
-  } catch (error) {
-    throw error
   }
 
-  const dateString = new Date().toISOString().split(/[:.-]/).join('_');
   // save pdf file
+  const dateString = new Date().toISOString().split(/[:.-]/).join('_');
   const pdfFile = new Promise((resolve, reject) => {
     const savePdfPath = path.join(
       app.getPath('userData'),
@@ -84,12 +76,8 @@ const pdfGenerator = async (latexString, type, id) => {
   );
   const texFile = fsPromises.writeFile(saveTexPath, latexString);
 
-  try {
-    const returnArray = await Promise.all([pdfFile, texFile]);
-    return returnArray[0];
-  } catch (error) {
-    throw error;
-  }
+  const returnArray = await Promise.all([pdfFile, texFile]);
+  return returnArray[0];
 };
 
 const pdfGeneratorHandler = async (event, type, args) => {
@@ -102,4 +90,4 @@ const pdfGeneratorHandler = async (event, type, args) => {
   return pdfGenerator(latexString, type, args.id);
 };
 
-module.exports = pdfGeneratorHandler;
+module.exports = { pdfGeneratorHandler };