+++ src/main/main.js
@@ -12,12 +12,12 @@ const url = require('url');
 const fs = require('fs');
 const isDev = require('electron-is-dev');
 const { databaseInit, databaseHandler } = require('./commands/database');
-const pdfGeneratorHandler = require('./commands/pdfGenerator/pdfGenerator');
-const { saveJobDescription } = require('./commands/saveJobDescription');
+const { pdfGeneratorHandler } = require('./commands/pdfGenerator/pdfGenerator');
 const { getUserDataPath } = require('./commands/getPaths');
 const { exportToCsv } = require('./commands/export');
-const { saveCV } = require('./commands/saveCV');
+const { savePdf } = require('./commands/savePdf');
 const { configManagement } = require('./commands/settings');
+const { openFolder } = require('./commands/openFileExplorer');
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 // eslint-disable-next-line global-require
@@ -159,6 +159,6 @@ ipcMain.handle('get-user-data-path', getUserDataPath);
 ipcMain.handle('settings', configManagement);
 ipcMain.handle('generate-pdf', pdfGeneratorHandler);
 ipcMain.handle('database', databaseHandler);
-ipcMain.handle('save-job-description', saveJobDescription);
-ipcMain.handle('save-cv', saveCV);
+ipcMain.handle('save-pdf', savePdf);
 ipcMain.handle('export-to-csv', exportToCsv);
+ipcMain.handle('open-folder', openFolder);