+++ src/renderer/pages/ApplicationPage/ApplicationDetails.jsx
@@ -1,47 +1,37 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useContext } from 'react';
 import {
   EmojiFrown,
   EmojiNeutral,
   EmojiSmile,
   EmojiSunglasses,
-  Tag,
   ArrowUpRightSquare,
 } from 'react-bootstrap-icons';
 import { updateDatabaseEntry } from '../../util/CRUD';
-import { Button, Selector } from '../../components/microComponents';
+import { Selector } from '../../components/microComponents';
 import { GlobalContext } from '../../context/GlobalContext';
 
 const ApplicationDetails = ({ id, appsData, setAppsData }) => {
   const appDetails = appsData.find((elem) => elem.id === id);
 
-  const { seasonValues, currentSeason } = useContext(GlobalContext);
-
+  const { seasonValues } = useContext(GlobalContext);
   const updateValue = (newValue, field) => {
-    if (field === 'company') {
-      const newAppsData = appsData.map((elem) => {
-        if (elem.id === id) return { ...elem, company: newValue };
+    let newAppsData = [...appsData];
+    const acceptableFields = [
+      'company',
+      'role',
+      'status',
+      'link',
+      'priority',
+      'location',
+      'deadline',
+    ];
+    if (acceptableFields.includes(field)) {
+      newAppsData = newAppsData.map((elem) => {
+        if (elem.id === id) return { ...elem, [field]: newValue };
         return elem;
       });
-      setAppsData(newAppsData);
-    } else if (field === 'role') {
-      const newAppsData = appsData.map((elem) => {
-        if (elem.id === id) return { ...elem, role: newValue };
-        return elem;
-      });
-      setAppsData(newAppsData);
-    } else if (field === 'status') {
-      const newAppsData = appsData.map((elem) => {
-        if (elem.id === id) return { ...elem, status: newValue };
-        return elem;
-      });
-      setAppsData(newAppsData);
-    } else if (field === 'link') {
-      const newAppsData = appsData.map((elem) => {
-        if (elem.id === id) return { ...elem, link: newValue };
-        return elem;
-      });
-      setAppsData(newAppsData);
     }
+    setAppsData(newAppsData);
   };
 
   const saveValue = (field) => {
@@ -109,8 +99,24 @@ const ApplicationDetails = ({ id, appsData, setAppsData }) => {
               onChange={(event) => updateValue(event.target.value, 'role')}
               onBlur={() => saveValue('role')}
             />
+            <input
+              className='bg-inherit p-1 text-base text-gray-700 outline-blue-500 hover:outline-blue-500'
+              type='text'
+              title='Location'
+              placeholder='1600 Pennsylvania Ave., NW Washington, DC 20500'
+              value={appDetails.location}
+              onChange={(event) => updateValue(event.target.value, 'location')}
+              onBlur={() => saveValue('location')}
+            />
+            <input
+              className='bg-inherit p-1 text-base text-gray-700 outline-blue-500 hover:outline-blue-500'
+              type='date'
+              title='Deadline'
+              value={appDetails.deadline}
+              onChange={(event) => updateValue(event.target.value, 'deadline')}
+              onBlur={() => saveValue('deadline')}
+            />
             <div className='flex items-center gap-x-2'>
-              {' '}
               <input
                 className='bg-inherit p-1 text-base text-gray-700 outline-blue-500'
                 type='text'
@@ -129,9 +135,8 @@ const ApplicationDetails = ({ id, appsData, setAppsData }) => {
                 <ArrowUpRightSquare className='mr-1' />
                 Open
               </a>
-              {/* <Button value="Open" Icon={ArrowUpRightSquare} color="purple" /> */}
             </div>
-            <div className='my-1'>
+            <div>
               <Selector
                 options={[
                   { k: 'To apply', v: 'To apply' },
@@ -143,12 +148,17 @@ const ApplicationDetails = ({ id, appsData, setAppsData }) => {
                   { k: 'Offer', v: 'Offer' },
                 ]}
                 title='Status'
+                style={{
+                  backgroundColor: 'inherit',
+                  borderColor: 'transparent',
+                  padding: '0.25rem',
+                }}
                 selected={appDetails.status}
                 onChange={(event) => updateValue(event.target.value, 'status')}
                 onBlur={() => saveValue('status')}
               />
             </div>
-            <div className='my-1'>
+            <div>
               <Selector
                 options={[
                   { k: 'low', v: 'Low' },
@@ -156,6 +166,12 @@ const ApplicationDetails = ({ id, appsData, setAppsData }) => {
                   { k: 'high', v: 'High' },
                 ]}
                 title='Priority'
+                style={{
+                  backgroundColor: 'inherit',
+                  borderColor: 'transparent',
+                  padding: '0.25rem',
+                }}
+                className='border-transparent bg-inherit p-0'
                 selected={appDetails.priority}
                 onChange={(event) =>
                   updateValue(event.target.value, 'priority')