+++ src/renderer/pages/ApplicationPage/CVSection/CVConstructorSection/CVConstructorSection.jsx
@@ -21,6 +21,7 @@ import schema from '../../../../constants/template2_schema';
 import 'tw-elements/dist/src/js/index';
 import { toast } from 'react-hot-toast';
 import { Button } from '../../../../components/microComponents';
+import { genericErrorNotification } from '../../../../components/Notifications';
 
 const CVConstructorSection = ({ id }) => {
   const { setAppsData, userPath } = useContext(GlobalContext);
@@ -46,7 +47,7 @@ const CVConstructorSection = ({ id }) => {
           [id, deleteComponentID],
           ({ error }) => {
             if (error) console.error(error);
-            setNoElementsClicked(noElementsClicked + 1);
+            else setNoElementsClicked(noElementsClicked + 1);
           }
         );
       }
@@ -57,7 +58,7 @@ const CVConstructorSection = ({ id }) => {
           [id, deleteComponentID],
           ({ error }) => {
             if (error) console.error(error);
-            setNoElementsClicked(noElementsClicked + 1);
+            else setNoElementsClicked(noElementsClicked + 1);
           }
         );
       }
@@ -86,14 +87,16 @@ const CVConstructorSection = ({ id }) => {
       ({ error }) => {
         if (error) console.error(error);
         // TODO add a check to make sure that component deleted is not referenced by another application component
-        deleteDatabaseEntry(
-          'DELETE FROM cv_components WHERE id = ?',
-          [componentId],
-          ({ error }) => {
-            if (error) console.error(error);
-            setNoElementsClicked(noElementsClicked + 1);
-          }
-        );
+        else {
+          deleteDatabaseEntry(
+            'DELETE FROM cv_components WHERE id = ?',
+            [componentId],
+            ({ error }) => {
+              if (error) console.error(error);
+              else setNoElementsClicked(noElementsClicked + 1);
+            }
+          );
+        }
       }
     );
   };
@@ -136,13 +139,15 @@ const CVConstructorSection = ({ id }) => {
   };
 
   const addCVSectionBuilderHandler = (sectionObj, sectionDesc) => {
+    const dateString = new Date().toISOString();
     createDatabaseEntry(
-      'INSERT INTO cv_components (cv_component_section, cv_component_text, cv_component_description, date_created) VALUES (?,?,?,?)',
+      'INSERT INTO cv_components (cv_component_section, cv_component_text, cv_component_description, date_created, date_modified) VALUES (?,?,?,?,?)',
       [
         sectionObj.section,
         JSON.stringify(sectionObj[sectionObj.section], null, 2),
         sectionDesc,
-        new Date().toISOString(),
+        dateString,
+        dateString,
       ],
       ({ error }) => {
         if (error) console.error(error);
@@ -170,7 +175,9 @@ const CVConstructorSection = ({ id }) => {
   const generatePdf = () => {
     if (elements.filter((elem) => elem.application_id).length === 0) {
       console.error('Select CV elements before generating document!');
-      toast.error('Error: Select CV elements before generating document');
+      genericErrorNotification(
+        'Error: Select CV elements before generating document'
+      );
       return;
     }
     /**
@@ -198,7 +205,6 @@ const CVConstructorSection = ({ id }) => {
             );
           }
         );
-        console.log('New CV PDF url: ', relativeCVUrl);
       })
       .catch((error) => {
         console.error(`PDF error: ${error}`);
@@ -215,10 +221,9 @@ const CVConstructorSection = ({ id }) => {
                 <Button
                   Icon={Folder2Open}
                   value='Open'
-                  onClick={openFileExplorer(`${userPath}${savePath}`)}
+                  onClick={() => openFileExplorer(`${userPath}${savePath}`)}
                 />
               </span>
-              {/* <button onClick><XLg /></button> */}
             </div>
           );
         },
@@ -255,7 +260,9 @@ const CVConstructorSection = ({ id }) => {
       : name ?? text ?? institution ?? organization ?? title ?? language;
   };
 
-  const openFileExplorer = (path) => {};
+  const openFileExplorer = (path) => {
+    window.electron.openFolder(path);
+  };
 
   return (
     <div className='mb-2'>